+++ RoutingController/RoutingControllElements/Link.cs
@@ -6,9 +6,9 @@ namespace RoutingController.RoutingControllElements
     {
         public string SourceId { get; set; }
         public string DestinationId { get; set; }
-        public double Weight { get; set; }
+        public int Weight { get; set; }
 
-        public Link(string sourceId, string destinationId, double weight)
+        public Link(string sourceId, string destinationId, int weight)
         {
             this.SourceId = sourceId;
             this.DestinationId = destinationId;
